agents/requirements.txt
google-adk
fastapi
uvicorn
pydantic
python-dotenv sr